+++ src/pcare/referensi.ts
@@ -120,33 +120,6 @@ export class Referensi extends PCareBaseApi {
 			method: 'GET'
 		});
 	}
-
-	/**
-	 * Get Data Referensi Spesialis
-	 */
-	async spesialis() {
-		return this.send<{ count: number; list: SpesialisResult[] }>({
-			name: `${this.name} Data Referensi Spesialis`,
-			path: '/spesialis',
-			method: 'GET'
-		});
-	}
-
-	/**
-	 * Get Data Referensi Sub Spesialis
-	 * @param params params{kode}
-	 * @returns
-	 */
-	async subSpesialis(params: {
-		/** Kode Spesialis */
-		kode: string;
-	}) {
-		return this.send<{ count: number; list: SubSpesialisResult[] }>({
-			name: `${this.name} Data Referensi Sub Spesialis`,
-			path: ['/spesialis/:kode/subspesialis', params],
-			method: 'GET'
-		});
-	}
 }
 
 interface PoliResult {
@@ -217,23 +190,4 @@ interface ProviderResult {
 
 	/** nama provider */
 	nmProvider: string;
-}
-
-interface SpesialisResult {
-	/** kode spesialis */
-	kdSpesialis: string;
-
-	/** nama spesialis */
-	nmSpesialis: string;
-}
-
-interface SubSpesialisResult {
-	/** kode sub spesialis */
-	kdSubSpesialis: string;
-
-	/** nama sub spesialis */
-	nmSubSpesialis: string;
-
-	/** kode poli rujuk  */
-	kdPoliRujuk: string;
 }
